+++ crates/quackdb-internal/src/handles/mod.rs
@@ -14,3 +14,5 @@ mod logical_type;
 pub use logical_type::*;
 mod table_function;
 pub use table_function::*;
+mod replacement_scan;
+pub use replacement_scan::*;