+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `ToDuckDb` types
 - `chrono` support
+- Minimal replacement scan support
 
 ### Changed
 - Internal bind/append API now uses duckdb types directly